requirements.txt
Click==8.1.3
dash==2.7.1
Flask==2.2.2
Flask-Compress==1.13
future==0.18.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
pytz==2022.7
retrying==1.3.4
six==1.16.0
Werkzeug==2.2.2
gunicorn>=20.1.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
pyyaml==6.0
plotly==5.11.0
requests==2.28.1
python-dateutil==2.8.2
fredapi==0.5.0
nasdaq-data-link==1.0.4
dash-extensions==0.1.8
pandas-datareader==0.10.0